+++ lib/src/local_image_stub.dart
@@ -17,6 +17,8 @@ Widget buildLocalRasterImage({
   required Color? color,
   required BlendMode? blendMode,
   required Alignment alignment,
+  required String? semanticLabel,
+  required bool excludeFromSemantics,
   required ImageErrorWidgetBuilder? errorBuilder,
 }) {
   return buildLocalRasterPathImage(
@@ -27,6 +29,8 @@ Widget buildLocalRasterImage({
     color: color,
     blendMode: blendMode,
     alignment: alignment,
+    semanticLabel: semanticLabel,
+    excludeFromSemantics: excludeFromSemantics,
     errorBuilder: errorBuilder,
   );
 }
@@ -39,6 +43,8 @@ Widget buildLocalRasterPathImage({
   required Color? color,
   required BlendMode? blendMode,
   required Alignment alignment,
+  required String? semanticLabel,
+  required bool excludeFromSemantics,
   required ImageErrorWidgetBuilder? errorBuilder,
 }) {
   return Image.network(
@@ -49,6 +55,8 @@ Widget buildLocalRasterPathImage({
     color: color,
     colorBlendMode: blendMode,
     alignment: alignment,
+    semanticLabel: semanticLabel,
+    excludeFromSemantics: excludeFromSemantics,
     errorBuilder: errorBuilder,
   );
 }
@@ -59,6 +67,8 @@ Widget buildLocalSvgImage({
   required double? width,
   required BoxFit? fit,
   required Alignment alignment,
+  required String? semanticsLabel,
+  required bool excludeFromSemantics,
   required ColorFilter? colorFilter,
   required Widget Function(BuildContext, Object, StackTrace?)? errorBuilder,
 }) {
@@ -68,6 +78,8 @@ Widget buildLocalSvgImage({
     width: width,
     fit: fit,
     alignment: alignment,
+    semanticsLabel: semanticsLabel,
+    excludeFromSemantics: excludeFromSemantics,
     colorFilter: colorFilter,
     errorBuilder: errorBuilder,
   );
@@ -79,6 +91,8 @@ Widget buildLocalSvgPathImage({
   required double? width,
   required BoxFit? fit,
   required Alignment alignment,
+  required String? semanticsLabel,
+  required bool excludeFromSemantics,
   required ColorFilter? colorFilter,
   required Widget Function(BuildContext, Object, StackTrace?)? errorBuilder,
 }) {
@@ -88,6 +102,8 @@ Widget buildLocalSvgPathImage({
     width: width,
     fit: fit ?? BoxFit.contain,
     alignment: alignment,
+    semanticsLabel: semanticsLabel,
+    excludeFromSemantics: excludeFromSemantics,
     colorFilter: colorFilter,
     errorBuilder: errorBuilder,
   );